提交星球聚合

+++ only4-KSP-cap4j-domain/src/main/java/com/only4/domain/aggregates/star/factory/StarFactory.java
@@ -39,7 +39,7 @@ public class StarFactory implements AggregateFactory<StarFactory.Payload, Star> 
     @AllArgsConstructor
     public static class Payload implements AggregatePayload<Star> {
         String name;
-        Long amount;
+        Integer amount;
         String description;
 
     }
